% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractionBigQuery.R
\name{extractionBigQuery}
\alias{extractionBigQuery}
\title{Extraction des données de Big Query}
\usage{
extractionBigQuery(query, projet_id, bdd_nom)
}
\arguments{
\item{query}{Requête SQL à exécuter sous Big Query}

\item{projet_id}{Identifiant Big Queru du projet}

\item{bdd_nom}{Nom de la base de données finale}
}
\value{
Un \code{data.frame} des données extraites de Big Query
}
\description{
À partir de R, envoi d'une requête SQL à Big Query, puis extraction et récupération des données sous R
}
\details{
La fonction "extractionBigQuery" envoie une requête SQL à exécuter sous Big Query pour extraire les données.
Ces dernières sont stockées dans une table Big Query, puis transférées sous Google Cloud Storage. À partir
de cette emplacement, la fonction télécharge les données au format .csv dans le répertoire courant. La fonction
importe ces données sous R et les enregistre au format .rds. Enfin, la fonction concatène l'ensemble des données
dans un seul data frame.
}
\author{
Yoann Pouëssel
}
