% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofmBigQuery.R
\docType{package}
\name{ofmBigQuery}
\alias{ofmBigQuery}
\alias{ofmBigQuery-package}
\title{Package "maison" permettant d'interagir avec Big Query}
\description{
Package contenant la fonction "extractionBigQuery".
}
\author{
Yoann Pouëssel
}
